refactor(main): break the main.dart / router.dart import cycle
router.dart imported main.dart for ProviderInitializer and BeerFestivalHome
while main.dart imported router.dart for appRouter, so main.dart was not an
entry point but a widget library that happened to contain main(), and
neither file could be read or tested without the other.

Move both widgets to lib/widgets/ and export them from the barrel. They are
copied verbatim — the only edit is one comment pointer.

globalRoutes moves from router.dart to utils/navigation_helpers.dart rather
than staying put as the issue suggested: the redirect handler that reads it
now lives under lib/widgets/, and router.dart imports the widgets barrel, so
leaving it in router.dart would simply re-form the cycle one file over.
navigation_helpers.dart has no project imports of its own and already owns
the route-path builders, so it is the natural cycle-free home.

main.dart is now main() + isTransientFontLoadError + BeerFestivalApp, 104
lines instead of 440. Nothing under lib/ imports it any more.

Removing the moved code left six dead imports in main.dart, which the
newly-fatal analyzer from #524 caught immediately.

Follow-up not done here, to keep this a pure move: the BeerFestivalHome and
ProviderInitializer tests still live in test/main_test.dart rather than
mirroring the new lib/widgets/ layout.

Fixes #527
