+++ lib/widgets/widgets.dart
@@ -1,3 +1,4 @@
+export 'beer_festival_home.dart';
 export 'brewery_hero_panel.dart';
 export 'collapsing_detail_app_bar.dart';
 export 'drink_card.dart';
@@ -14,6 +15,7 @@ export 'highlighted_text.dart';
 export 'info_chip.dart';
 export 'overflow_menu.dart';
 export 'page_title.dart';
+export 'provider_initializer.dart';
 export 'section_header.dart';
 export 'sheet_handle.dart';
 export 'star_rating.dart';
