+++ test/main_test.dart
@@ -3,6 +3,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:cambridge_beer_festival/main.dart';
 import 'package:cambridge_beer_festival/screens/screens.dart';
+import 'package:cambridge_beer_festival/widgets/widgets.dart';
 import 'package:cambridge_beer_festival/providers/beer_provider.dart';
 import 'package:cambridge_beer_festival/services/services.dart';
 import 'package:cambridge_beer_festival/models/models.dart';
